Clk: Enable i2c1 to manager i2c1 clk rate during sys runtime
update as below:
Some clk is disabled and unprepared, but clk rate is not 0.
For example, i2c1 is disabled in dts file, but its status is
enabled and unprepared, so rate/voltage will be initialized by DVFS.
This will impact vdd-core's voltage.

So we enable i2c1 module to manager clk rate.

Bug 200493082

Change-Id: I788130042548073582f93c6840d3bcea8bca0189
Signed-off-by: Taylor Xiao <txiao@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/2031939
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Terry Wang <terwang@nvidia.com>
Tested-by: Terry Wang <terwang@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: David Dastous St Hilaire <ddastoussthi@nvidia.com>
Reviewed-by: Vinayak Pane <vpane@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>